generic_linux.pin for pi; fallback to rpi.gpio if libgpiod unavailable
