+++ src/adafruit_blinka/microcontroller/raspi_23/pin.py
@@ -1,62 +1,9 @@
-import RPi.GPIO as GPIO
-GPIO.setmode(GPIO.BCM)    # Use BCM pins D4 = GPIO #4
-GPIO.setwarnings(False)   # shh!
-
-# Pins dont exist in CPython so...lets make our own!
-class Pin:
-    IN = 0
-    OUT = 1
-    LOW = 0
-    HIGH = 1
-    PULL_NONE = 0
-    PULL_UP = 1
-    PULL_DOWN = 2
-    
-    id = None
-    _value = LOW
-    _mode = IN
-    
-    def __init__(self, bcm_number):
-        self.id = bcm_number
-
-    def __repr__(self):
-        return str(self.id)
-
-    def __eq__(self, other):
-        return self.id == other
-
-    def init(self, mode=IN, pull=None):
-        if mode != None:
-            if mode == self.IN:
-                self._mode = self.IN
-                GPIO.setup(self.id, GPIO.IN)
-            elif mode == self.OUT:
-                self._mode = self.OUT
-                GPIO.setup(self.id, GPIO.OUT)
-            else:
-                raise RuntimeError("Invalid mode for pin: %s" % self.id)
-        if pull != None:
-            if self._mode != self.IN:
-                raise RuntimeError("Cannot set pull resistor on output")
-            if pull == self.PULL_UP:
-                GPIO.setup(self.id, GPIO.IN, pull_up_down=GPIO.PUD_UP)
-            elif pull == self.PULL_DOWN:
-                GPIO.setup(self.id, GPIO.IN, pull_up_down=GPIO.PUD_DOWN)
-            else:
-                raise RuntimeError("Invalid pull for pin: %s" % self.id)       
-
-    def value(self, val=None):
-        if val != None:
-            if val == self.LOW:
-                self._value = val
-                GPIO.output(self.id, val)
-            elif val == self.HIGH:
-                self._value = val
-                GPIO.output(self.id, val)
-            else:
-                raise RuntimeError("Invalid value for pin")
-        else:
-            return GPIO.input(self.id)
+try:
+    # Start by seeing if we can support the libgpiod-based generic linux driver:
+    from adafruit_blinka.microcontroller.generic_linux.pin import Pin
+except ImportError:
+    # Failing that, fall back to the RPi.GPIO-based version:
+    from adafruit_blinka.microcontroller.raspi_23.pin_rpi_gpio import Pin
 
 D2 = Pin(2)
 SDA = Pin(2)
@@ -115,4 +62,3 @@ uartPorts = (
 i2cPorts = (
     (1, SCL, SDA),
 )
-
